pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "discord-py[test]>=2.4.0",
    "flask>=3.1.0",
    "pytest-asyncio>=0.25.3",
    "pytest>=8.3.4",
    "python-dotenv>=1.0.1",
]
